refactored compressFile into separate module
