+++ src/brotli-plugin.js
@@ -1,11 +1,10 @@
 'use strict'
 
-const brotli = require('brotli')
-const fs = require('fs')
+
 const glob = require('glob')
-const mkdirp = require('mkdirp')
 const path = require('path')
 const util = require('util')
+const compressFile = require('./compressFile')
 
 const defaultOptions = {
   extensions: ['css', 'js'],
@@ -13,24 +12,6 @@ const defaultOptions = {
 }
 
 const globAsync = util.promisify(glob)
-const mkdirpAsync = util.promisify(mkdirp)
-const readFileAsync = util.promisify(fs.readFile)
-const writeFileAsync = util.promisify(fs.writeFile)
-
-async function compressFile (file, pluginOptions = {}) {
-  // brotli compress the asset to a new file with the .br extension
-  const fileBasePath = path.join(process.cwd(), 'public')
-  const srcFileName = path.join(fileBasePath, file)
-  const content = await readFileAsync(srcFileName)
-  const compressed = await brotli.compress(content)
-
-  const destFilePath = (pluginOptions.path) ? path.join(fileBasePath, pluginOptions.path) : fileBasePath
-  const destFileName = path.join(destFilePath, file) + '.br'
-  const destFileDirname = path.dirname(destFileName)
-
-  await mkdirpAsync(destFileDirname)
-  await writeFileAsync(destFileName, compressed)
-}
 
 async function onPostBuild (args, pluginOptions) {
   const options = { ...defaultOptions, ...pluginOptions }
